backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
pymongo==4.6.2
requests==2.31.0
httpx==0.24.1
pytest==8.0.2
Jinja2==3.1.3
starlette==0.36.3
isort==5.13.2
black==24.10.0